+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-- Nested-field returning changed line index
-- Nested-field enumerating label
+- `QasNestedFields` returning changed line index.
+- `QasNestedFields` enumerating label.
 
 ## 2.5.1 - 2021-07-23